canary: try to match commit to PR if not found in env
